+++ dwds/lib/src/utilities/wrapped_service.dart
@@ -19,7 +19,39 @@ export 'package:vm_service/vm_service.dart'
         InstanceRef,
         FuncRef,
         FieldRef,
-        CodeRef;
+        CodeRef,
+        Event,
+        Stack;
+
+class Event extends vm_service.Event {
+  Event._();
+
+  factory Event({
+    @required String kind,
+    @required int timestamp,
+  }) {
+    var event = Event._();
+    event
+      ..kind = kind
+      ..timestamp = timestamp;
+    return event;
+  }
+}
+
+class Stack extends vm_service.Stack {
+  Stack._();
+
+  factory Stack({
+    @required List<vm_service.Frame> frames,
+    @required List<vm_service.Message> messages,
+  }) {
+    var stack = Stack._();
+    stack
+      ..frames = frames
+      ..messages = messages;
+    return stack;
+  }
+}
 
 class FieldRef extends vm_service.FieldRef {
   FieldRef._();